Cleaned up sections of the website
