--- src/Hero.css
@@ -1,15 +1,0 @@
-#logo, #logoBackground {
-    position: absolute;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-}
-
-#logo {
-    width: 50%;
-}
-
-#logoBackground {
-    width: 60%;
-    z-index: -1;
-}
