+++ src/components/Footer.css
@@ -1,6 +1,4 @@
 #footer {
-    position: absolute;
-    top: 880%;
     display: flex;
     flex-direction: column;
     align-items: center;
