+++ src/components/Nav.js
@@ -1,29 +1,39 @@
-import './Nav.css';
-import '../../src/Global.css';
-import React from 'react';
+import "./Nav.css";
+import "../Global.css";
+import React from "react";
 
 const Nav = () => {
   return (
     <div>
       <ul id="navigation">
         <li key="hero">
-          <a id="menuHero" className="menu" href="#hero">home</a>
+          <a id="menuHero" className="menu" href="#hero">
+            home
+          </a>
         </li>
         <li key="about">
-          <a id='menuAbout' className="menu" href="#about">about</a>
+          <a id="menuAbout" className="menu" href="#about">
+            about
+          </a>
         </li>
         <li key="projects">
-          <a id='menuProject' className="menu" href="#projects">projects</a>
+          <a id="menuProject" className="menu" href="#projects">
+            projects
+          </a>
         </li>
         <li key="contact">
-          <a id='menuProject' className="menu" href="#contact">contact</a>
+          <a id="menuProject" className="menu" href="#contact">
+            contact
+          </a>
         </li>
         <li key="resume">
-          <a id="menuResume" className='menu' href="resume.pdf" target="_blank">résumé</a>
+          <a id="menuResume" className="menu" href="resume.pdf" target="_blank">
+            résumé
+          </a>
         </li>
       </ul>
     </div>
   );
-}
+};
 
 export default Nav;
