+++ src/components/Projects.css
@@ -1,6 +1,4 @@
 #projects {
-    position: absolute;
-    top: 225%;
     display: flex;
     flex-direction: column;
     align-items: center;
