+++ src/index.js
@@ -1,14 +1,13 @@
-import './index.css';
-import React from 'react';
-import ReactDOM from 'react-dom';
-import { BrowserRouter as Router} from "react-router-dom"; 
-import Nav from './components/Nav';
-import SideNav from './components/SideNav';
-import Hero from './Hero';
-import About from './About';
-import Projects from './Projects';
-import Contact from './Contact';
-import Footer from './Footer';
+import "./index.css";
+import React from "react";
+import ReactDOM from "react-dom";
+import { BrowserRouter as Router } from "react-router-dom";
+import Nav from "./components/Nav";
+import SideNav from "./components/SideNav";
+import Hero from "./components/Hero";
+import About from "./components/About";
+import Projects from "./components/Projects";
+import Footer from "./components/Footer";
 
 ReactDOM.render(
   <Router>
@@ -17,8 +16,7 @@ ReactDOM.render(
     <Hero />
     <About />
     <Projects />
-    <Contact />
     <Footer />
   </Router>,
-  document.getElementById('root')
+  document.getElementById("root")
 );
